# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "The JSON interface to a HyperUplink board, exposing the same\ncategories, forums, topics and administrative surfaces the web\ninterface serves, with every response shaped by the permissions\nof the user the API key belongs to.\n\nAuthentication happens with an API key that a signed-in user\nmints under Account -\u003e API on the board itself, and the secret\nit hands out once is sent either as a bearer token in the\nAuthorization header or verbatim in the X-API-Key header, so a\nrequest that carries neither is answered with 401 before it\never reaches a controller.",
        "title": "HyperUplink API",
        "contact": {
            "name": "HyperUplink",
            "url": "https://hyperup.link"
        },
        "license": {
            "name": "SEGV",
            "url": "https://xn--gckvb8fzb.com/segv/"
        },
        "version": "{{.Version}}"
    },
    "basePath": "/",
    "paths": {
...
        "/_{categories}/{forums}/{topics}/poll": {
            "post": {
                "description": "A vote is final and can only be cast while the poll is still open,\nand the results are what the topic hands back afterwards.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "board"
                ],
                "summary": "Vote in a topic's poll",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The category slug",
                        "name": "categories",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The forum slug",
                        "name": "forums",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The topic slug",
                        "name": "topics",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The option to vote for",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/topics.TopicPollVoteBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/request.StatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/request.ValidationErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            }
        },
...
        "/new": {
            "get": {
                "description": "Hands back the forums the caller may post in, and, when a forum,\ntopic or reply is named in the query, the post that is being\nanswered.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "board"
                ],
                "summary": "Show what a new post may be written into",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The slug of the forum to post in",
                        "name": "forum",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The slug of the topic to reply to",
                        "name": "topic",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The identifier of the reply to quote",
                        "name": "reply",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/newpost.FormView"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            },
            "post": {
                "description": "Creates a topic in the given forum, optionally with a poll and with\nattachments that were uploaded beforehand, and answers with the slugs\nthe new topic can be read under.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "board"
                ],
                "summary": "Start a topic",
                "parameters": [
                    {
                        "description": "The topic to create",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/newpost.NewCreateBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "category_slug": {
                                    "type": "string"
                                },
                                "forum_slug": {
                                    "type": "string"
                                },
                                "id": {
                                    "type": "string"
                                },
                                "slug": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/request.ValidationErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "category.Category": {
            "type": "object",
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "deleted_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                }
            }
        },
...
        "forum.Forum": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "created_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "deleted_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                }
            }
        },
...
        "github_com_google_uuid.NullUUID": {
            "type": "object",
            "properties": {
                "uuid": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if UUID is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "newpost.CategoryWithForums": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/category.Category"
                },
                "forums": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/forum.Forum"
                    }
                }
            }
        },
        "newpost.FormView": {
            "type": "object",
            "properties": {
                "allow_poll": {
                    "type": "boolean"
                },
                "categories_forums": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/newpost.CategoryWithForums"
                    }
                },
                "forum": {
                    "$ref": "#/definitions/vforum.VForum"
                },
                "poll_options_max": {
                    "type": "integer"
                },
                "reply": {
                    "$ref": "#/definitions/reply.Reply"
                },
                "topic": {
                    "$ref": "#/definitions/vtopic.VTopic"
                }
            }
        },
        "newpost.NewCreateBody": {
            "type": "object",
            "required": [
                "forum_id",
                "name",
                "text"
            ],
            "properties": {
                "attachment_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "forum_id": {
                    "type": "string"
                },
                "kind": {
                    "type": "string",
                    "enum": [
                        "regular",
                        "poll"
                    ]
                },
                "name": {
                    "type": "string",
                    "maxLength": 78,
                    "minLength": 1
                },
                "poll_ends_at": {
                    "type": "string"
                },
                "poll_options": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "text": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
...
        "pgtype.InfinityModifier": {
            "type": "integer",
            "format": "int32",
            "enum": [
                1,
                0,
                -1
            ],
            "x-enum-varnames": [
                "Infinity",
                "Finite",
                "NegativeInfinity"
            ]
        },
...
        "pgtype.Timestamp": {
            "type": "object",
            "properties": {
                "infinityModifier": {
                    "$ref": "#/definitions/pgtype.InfinityModifier"
                },
                "time": {
                    "description": "Time zone will be ignored when encoding to PostgreSQL.",
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
...
        "reply.Reply": {
            "type": "object",
            "properties": {
                "attachment_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "author_id": {
                    "type": "string"
                },
                "created_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "deleted_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "html": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "moderated_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "reply_id": {
                    "$ref": "#/definitions/github_com_google_uuid.NullUUID"
                },
                "short_id": {
                    "type": "string"
                },
                "spammed_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "text": {
                    "type": "string"
                },
                "topic_id": {
                    "type": "string"
                },
                "updated_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                }
            }
        },
...
        "request.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "forbidden"
                }
            }
        },
        "request.StatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "request.ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "validation"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "topic.Kind": {
            "type": "string",
            "enum": [
                "regular",
                "question",
                "poll",
                "rsvp"
            ],
            "x-enum-varnames": [
                "Regular",
                "Question",
                "Poll",
                "RSVP"
            ]
        },
...
        "topics.TopicPollVoteBody": {
            "type": "object",
            "properties": {
                "selection": {
                    "type": "integer"
                }
            }
        },
...
        "vforum.VForum": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "category_name": {
                    "type": "string"
                },
                "category_slug": {
                    "type": "string"
                },
                "created_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "deleted_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_reply_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "replies": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "topics": {
                    "type": "integer"
                },
                "updated_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                }
            }
        },
...
        "vtopic.VTopic": {
            "type": "object",
            "properties": {
                "anonymous": {
                    "type": "boolean"
                },
                "attachment_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "author_email": {
                    "type": "string"
                },
                "author_id": {
                    "type": "string"
                },
                "author_joined_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "author_member_of": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "author_profile_picture": {
                    "type": "string"
                },
                "author_role": {
                    "type": "string"
                },
                "author_signature_html": {
                    "type": "string"
                },
                "author_signature_text": {
                    "type": "string"
                },
                "author_username": {
                    "type": "string"
                },
                "category_name": {
                    "type": "string"
                },
                "category_slug": {
                    "type": "string"
                },
                "created_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "deleted_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "ended_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "forum_id": {
                    "type": "string"
                },
                "forum_name": {
                    "type": "string"
                },
                "forum_slug": {
                    "type": "string"
                },
                "html": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "kind": {
                    "$ref": "#/definitions/topic.Kind"
                },
                "last_reply_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "locked_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "moderated_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "name": {
                    "type": "string"
                },
                "pinned": {
                    "type": "boolean"
                },
                "poll_options": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "replies": {
                    "type": "integer"
                },
                "short_id": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "spammed_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "text": {
                    "type": "string"
                },
                "updated_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "APIKeyAuth": {
            "description": "The API key secret on its own.",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "The API key secret prefixed with \"Bearer \".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}